feat(setting-up-environment): make more approachable
When I first saw this page,

- it looked like a _huge_ table that extended off my screen
- it said I would need "several" tools, which exacerbated this affect
- it looked like there was no explanation of how to install them on this
  page, and that I would need to click through to each and follow its
  instructions
- when I clicked to "edit on GitHub", it brought me to a [an incomprehensible GitHub page](https://github.com/starknet-io/starknet-docs/edit/main/components/Starknet/modules/quick-start/pages/environment-setup.adoc). It took some guesswork and determination to figure out how to contribute to this doc!

This fixes the first three items, but I'm not sure how to address the
last one. Maybe just a comment, for now? Even when you clone the repo,
if you're not used to adoc, then the cross-repo `include` is hard to
figure out.

I've also updated a couple other readability & formatting issues I came
across.